# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/balance": {
      "get": {
        "description": "Get current balance across all wallets and currencies",
        "operationId": "LedgerController_getBalance",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Balance retrieved successfully."
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Get balance",
        "tags": [
          "Ledger"
        ]
      }
    },
...
    "/api/v1/dashboard/api-keys": {
      "post": {
        "operationId": "DashboardController_createApiKey",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created successfully."
          },
          "400": {
            "description": "Validation failed or the request body/params were malformed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Create API key",
        "tags": [
          "Dashboard"
        ]
      },
      "get": {
        "operationId": "DashboardController_listApiKeys",
        "parameters": [],
        "responses": {
          "200": {
            "description": "API keys retrieved successfully."
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "List API keys",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/v1/dashboard/api-keys/{id}": {
      "patch": {
        "operationId": "DashboardController_updateApiKey",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "API key ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApiKeyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "API key updated successfully."
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource could not be found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Update API key label or IP whitelist",
        "tags": [
          "Dashboard"
        ]
      },
      "delete": {
        "operationId": "DashboardController_revokeApiKey",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "API key ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key revoked successfully."
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource could not be found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Revoke API key",
        "tags": [
          "Dashboard"
        ]
      }
    },
...
  },
  "info": {
    "title": "Dexxify API",
    "description": "Crypto Infrastructure API for Africa (Wallets, Payouts, Offramp, Onramp, KYC, KYB)",
    "version": "1.0",
    "contact": {
      "name": "Dexxify",
      "url": "https://www.dexxify.com",
      "email": "dexxifyhq@gmail.com"
    }
  },
...
  "servers": [
    {
      "url": "https://api.dexxify.com/api/v1",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "api-key": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "API key for /api/v1/* endpoints"
      }
    },
    "schemas": {
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "status": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Validation failed."
          },
          "errors": {
            "type": "object",
            "example": [
              "amount must not be less than 0"
            ]
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-17T12:00:00.000Z"
          }
        },
        "required": [
          "success",
          "status",
          "message",
          "timestamp"
        ]
      },
...
      "CreateApiKeyDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "description": "API key label",
            "example": "Mobile App Key"
          }
        }
      },
      "UpdateApiKeyDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "description": "API key label",
            "example": "Updated Mobile App Key"
          },
          "ip_whitelist": {
            "description": "IP whitelist",
            "example": [
              "192.168.1.1",
              "10.0.0.1"
            ],
            "type": "array",
            "items": {
              "type": "array"
            }
          }
        }
      },
...
    }
  }
}